Rename `ulp-riscv-hal` to `esp-ulp-riscv-hal`
