+++ .github/workflows/ci.yml
@@ -340,7 +340,7 @@ jobs:
       - name: rustdoc
         run: cd esp32h2-hal/ && cargo doc --features=eh1
 
-  ulp-riscv-hal:
+  esp-ulp-riscv-hal:
     runs-on: ubuntu-latest
 
     steps:
@@ -354,11 +354,11 @@ jobs:
 
       # Perform a full build initially to verify that the examples not only
       # build, but also link successfully.
-      - name: build ulp-riscv-hal (esp32s3)
-        run: cd ulp-riscv-hal/ && cargo +nightly build --release --features=esp32s3 --examples
+      - name: build esp-ulp-riscv-hal (esp32s3)
+        run: cd esp-ulp-riscv-hal/ && cargo +nightly build --release --features=esp32s3 --examples
       # Ensure documentation can be built
       - name: rustdoc
-        run: cd ulp-riscv-hal/ && cargo doc --features=esp32s3
+        run: cd esp-ulp-riscv-hal/ && cargo doc --features=esp32s3
 
   esp32s2-hal:
     runs-on: ubuntu-latest
@@ -377,10 +377,10 @@ jobs:
           components: rust-src
       - uses: Swatinem/rust-cache@v2
 
-      # build the ulp-riscv-hal examples first to make sure the examples which expect
+      # build the esp-ulp-riscv-hal examples first to make sure the examples which expect
       # the ELF files to be present will compile
-      - name: build ulp-riscv-hal prerequisites
-        run: cd ulp-riscv-hal/ && cargo +nightly build --release --features=esp32s2 --examples
+      - name: build esp-ulp-riscv-hal prerequisites
+        run: cd esp-ulp-riscv-hal/ && cargo +nightly build --release --features=esp32s2 --examples
 
       # Perform a full build initially to verify that the examples not only
       # build, but also link successfully.
@@ -452,10 +452,10 @@ jobs:
           components: rust-src
       - uses: Swatinem/rust-cache@v2
 
-      # build the ulp-riscv-hal examples first to make sure the examples which expect
+      # build the esp-ulp-riscv-hal examples first to make sure the examples which expect
       # the ELF files to be present will compile
-      - name: build ulp-riscv-hal prerequisites
-        run: cd ulp-riscv-hal/ && cargo +nightly build --release --features=esp32s3 --examples
+      - name: build esp-ulp-riscv-hal prerequisites
+        run: cd esp-ulp-riscv-hal/ && cargo +nightly build --release --features=esp32s3 --examples
 
       # Perform a full build initially to verify that the examples not only
       # build, but also link successfully.
@@ -596,10 +596,10 @@ jobs:
           components: rust-src
       - uses: Swatinem/rust-cache@v2
 
-      # build the ulp-riscv-hal examples first to make sure the examples which expect
+      # build the esp-ulp-riscv-hal examples first to make sure the examples which expect
       # the ELF files to be present will compile
-      - name: build ulp-riscv-hal prerequisites
-        run: cd ulp-riscv-hal/ && RUSTC_BOOTSTRAP=1 cargo +1.67 build --release --features=esp32s3 --examples
+      - name: build esp-ulp-riscv-hal prerequisites
+        run: cd esp-ulp-riscv-hal/ && RUSTC_BOOTSTRAP=1 cargo +1.67 build --release --features=esp32s3 --examples
 
       # Verify the MSRV for all Xtensa chips.
       - name: msrv (esp32-hal)
@@ -645,8 +645,8 @@ jobs:
         run: cargo +stable clippy --manifest-path=esp32c6-lp-hal/Cargo.toml -- -D warnings -A asm-sub-register
       - name: clippy (esp32h2-hal)
         run: cargo +stable clippy --manifest-path=esp32h2-hal/Cargo.toml -- -D warnings
-      - name: clippy (ulp-riscv-hal)
-        run: cargo +stable clippy --manifest-path=ulp-riscv-hal/Cargo.toml --features=esp32s3 -- -D warnings -A asm-sub-register
+      - name: clippy (esp-ulp-riscv-hal)
+        run: cargo +stable clippy --manifest-path=esp-ulp-riscv-hal/Cargo.toml --features=esp32s3 -- -D warnings -A asm-sub-register
 
   clippy-xtensa:
     runs-on: ubuntu-latest
@@ -706,5 +706,5 @@ jobs:
         run: cargo fmt --all --manifest-path=esp32s2-hal/Cargo.toml -- --check
       - name: rustfmt (esp32s3-hal)
         run: cargo fmt --all --manifest-path=esp32s3-hal/Cargo.toml -- --check
-      - name: rustfmt (ulp-riscv-hal)
-        run: cargo fmt --all --manifest-path=ulp-riscv-hal/Cargo.toml -- --check
+      - name: rustfmt (esp-ulp-riscv-hal)
+        run: cargo fmt --all --manifest-path=esp-ulp-riscv-hal/Cargo.toml -- --check
