+++ esp-hal-procmacros/src/lib.rs
@@ -127,11 +127,11 @@ fn get_hal_crate() -> (
     #[cfg(feature = "esp32s2")]
     let hal_crate = crate_name("esp32s2-hal");
     #[cfg(feature = "esp32s2-ulp")]
-    let hal_crate = crate_name("ulp-riscv-hal");
+    let hal_crate = crate_name("esp-ulp-riscv-hal");
     #[cfg(feature = "esp32s3")]
     let hal_crate = crate_name("esp32s3-hal");
     #[cfg(feature = "esp32s3-ulp")]
-    let hal_crate = crate_name("ulp-riscv-hal");
+    let hal_crate = crate_name("esp-ulp-riscv-hal");
 
     // Crate name:
     #[cfg(feature = "esp32")]
@@ -149,11 +149,11 @@ fn get_hal_crate() -> (
     #[cfg(feature = "esp32s2")]
     let hal_crate_name = Ident::new("esp32s2_hal", Span::call_site().into());
     #[cfg(feature = "esp32s2-ulp")]
-    let hal_crate_name = Ident::new("ulp_riscv_hal", Span::call_site().into());
+    let hal_crate_name = Ident::new("esp_ulp_riscv_hal", Span::call_site().into());
     #[cfg(feature = "esp32s3")]
     let hal_crate_name = Ident::new("esp32s3_hal", Span::call_site().into());
     #[cfg(feature = "esp32s3-ulp")]
-    let hal_crate_name = Ident::new("ulp_riscv_hal", Span::call_site().into());
+    let hal_crate_name = Ident::new("esp_ulp_riscv_hal", Span::call_site().into());
 
     (hal_crate, hal_crate_name)
 }
@@ -681,13 +681,13 @@ pub fn entry(args: TokenStream, input: TokenStream) -> TokenStream {
         crate_name("esp32c6-lp-hal").expect("esp32c6_lp_hal is present in `Cargo.toml`");
     #[cfg(any(feature = "esp32s2-ulp", feature = "esp32s3-ulp"))]
     let found_crate =
-        crate_name("ulp-riscv-hal").expect("ulp-riscv-hal is present in `Cargo.toml`");
+        crate_name("esp-ulp-riscv-hal").expect("esp-ulp-riscv-hal is present in `Cargo.toml`");
 
     let hal_crate = match found_crate {
         #[cfg(feature = "esp32c6-lp")]
         FoundCrate::Itself => quote!(esp32c6_lp_hal),
         #[cfg(any(feature = "esp32s2-ulp", feature = "esp32s3-ulp"))]
-        FoundCrate::Itself => quote!(ulp_riscv_hal),
+        FoundCrate::Itself => quote!(esp_ulp_riscv_hal),
         FoundCrate::Name(name) => {
             let ident = Ident::new(&name, Span::call_site());
             quote!( #ident::Something )
