+++ esp-ulp-riscv-hal/CHANGELOG.md
@@ -9,14 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add the `esp32c6-lp-hal` package (#714)
-- Add GPIO (output) and delay functionality to `esp32c6-lp-hal` (#715)
-- Add GPIO input support and implement additional `embedded-hal` output traits for the C6's LP core [#720]
+- Add the `ulp-riscv-hal` package (#840)
 
 ### Changed
+
+- Renamed to `esp-ulp-riscv-hal`
 
 ### Fixed
 
 ### Removed
 
-[Unreleased]: https://github.com/esp-rs/esp-hal/commits/main/esp32c6-lp-hal
+[Unreleased]: https://github.com/esp-rs/esp-hal/commits/main/esp-ulp-riscv-hal
