+++ esp-ulp-riscv-hal/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name         = "ulp-riscv-hal"
+name         = "esp-ulp-riscv-hal"
 version      = "0.1.0"
 edition      = "2021"
 rust-version = "1.67.0"
