+++ esp32s3-hal/examples/ulp_riscv_core_basic.rs
@@ -32,7 +32,7 @@ fn main() -> ! {
 
     // load code to LP core
     let lp_core_code = load_lp_code!(
-        "../ulp-riscv-hal/target/riscv32imc-unknown-none-elf/release/examples/blinky"
+        "../esp-ulp-riscv-hal/target/riscv32imc-unknown-none-elf/release/examples/blinky"
     );
 
     // start LP core
